requirements.txt
python-dotenv==1.1.1
numpy==2.3.3
websocket-client==1.8.0